feat: add exception handling for pre-students
The SAML of the university login has a possible affiliation value of 'pre-student'
We did not have proper error handling of this, making new members fear something
with their membership went wrong.
Added a sensible message letting the user know that they should wait until the
start of the academic year.